*Weight Truncated System with Prioritization*
a) added the weight system so that steering behaviors can be combined
b) added smoothing when rotating the object so that it face its heading direction. this is to prevent jittering when it collides with the wall.

The implementation simply saves X amount of heading vectors and override the oldest one when the list is full. The list is then used to average the heading vector to be used to transform the object.

Note: the code needs refactoring